Revert changes from PR #186
The change was maybe a bit to hars, When checking for mojang after minetools gives uuid null would result in unwanted delays / resource usage (as McLive stated). This could also be easily resolved if minetools gives status: ERR if things are not working well. Hench, future request should be done for minetools when the api is spitting out false status again. 

Making more lagg for big owners should not be the down payment when this issue is arising.