Yet another giant code rewrite (#529)
* Change renovate base branch to DEV (#478)

* Change renovate base branch to DEV (#480)

* Make more use of lombok and less too! :D

Ok the description may be a bit missleading. I removed lombok there where it can cause confusion.
imo lomboks @Setter should be used when you actually need to hotswap values sometimes.
And passing variables through the constructor if those have to be "final".
That is why i replaced some @Setters with @RequiredArgsConstructor

* Move quite a bit around and implement more API!

* People compile everything before they commit? pf who needs that

* Also add everything to bungeecord

* More rewriting of appliers

* And another one

* Make updater less complex

* Intellij cleanup

* Move properties into api package

* Clean more code

* Rewrite a lot of the applier code and likely also break backwards support

I am aware this will break a lot of stuff. There are chances to get this backwards working, but imo its not worth it. We had a major update already last month, but this is just very important.

* Cleanup bit stuff and rename skinstorage method

* Use sneaky throws

* And more rewriting!

* Change DEV version to 14.1.0-SNAPSHOT (again) (#484)

* And more refactoring!

* Also put bukkit into the boat!

* Chain instead of duplicating code!

* Cleanup and make version a part of ISRPlugin

* Create index page for website (#485)

* Create index.md

* Add sidebar

* And more!

* Less shame

* Better method names

* Optimize some code

* %skin for /skin set

* Set proper version comparator

* Fix build

* Revert updater

Co-authored-by: xknat <kevin.natte@gmail.com>