🔥 chore(api): remove unused code and plugins ✨ feat(api): add indra plugin and configure java versions The `net.kyori.indra` plugin is added to the `sr.base-logic.gradle.kts` file and configured to publish releases and snapshots to the CodeMC repository. The `javaVersions` block is added to the `sr.base-logic.gradle.kts` file to configure the target and minimum toolchain versions. Unused code and plugins are removed from the `api/build.gradle.kts` file. The `tasks.jar` and `tasks.shadowJar` blocks are added to the `bukkit/build.gradle.kts` file to include dependencies from the `folia` project. The `SRBukkitAdapter.java` file is updated to use the `SchedulerProvider` interface and the `ProviderSelector` class to select the appropriate implementation of the interface.
🚀 chore(gradle): add Folia dependency to multiver bukkit folia module
🐛 fix(EventBusImpl.java): remove usage of Lombok val
The Folia dependency is added to the multiver bukkit folia module to enable the use of the FoliaSchedulerProvider. The usage of Lombok val is removed from the EventBusImpl.java file to improve compatibility with older versions of Java.

🐛 fix(commands): remove unused import in SRCommand.java
✨ feat(commands): change dumpService method to return Optional<String>
🆕 feat(provider): add FeatureProvider interface
🆕 feat(provider): add ProviderSelector class
🆕 feat(velocity): add indra plugin and set java version to 11
The unused import in SRCommand.java has been removed. The dumpService method in SRCommand.java has been changed to return an Optional<String> instead of a val. The FeatureProvider interface and ProviderSelector class have been added. The indra plugin has been added to velocity's build.gradle.kts file and the java version has been set to 11.