Drop all support for Sponge
Sadly I was still the only person in the team holding on supporting Sponge, but I've also decided today that it's best. The sponge documentation hasn't been updated for newer version in a looong time and there are way too many APIs to support. And also noone is using sponge. Like 0. At least that much is what the statistics show. Due to stupid sponge rules we are not allowed to see how many people use SkinsRestorer without them going out of their way to go into server settings and setting it to true. There isn't even a command anymore to set the plugin to allow collecting stats. The only reason sponge was useful to us was that it supports forge, but we'd rather just directly support forge than support sponge.